fix delete

+++ server/resources/product.py
@@ -149,12 +149,14 @@ def delete(product_id):
     user_id = get_jwt_identity()
     print(f"user_id : {user_id}, product_id : {product_id}")
     p = ProductModel.query.filter_by(author_id = int(user_id), product_id = int(product_id)).first()
+    if not p:
+        raise error.DBNotFound('Product')
     
     favorites = FavoriteModel.query.filter_by(product_id=product_id).all()
     for favorite in favorites:
         rdb.session.delete(favorite)
+        
     rdb.session.delete(p)
-    
     rdb.session.commit()
 
     return jsonify({"status_code" : 200, "message":"Success"}), 200
